Rename the original session to session inner, and hide it
