+++ src/config/mod.rs
@@ -7,6 +7,7 @@ pub(crate) struct Config {
     pub ver: version::SshVersion,
     pub auth: auth::AuthInfo,
     pub algs: algorithm::AlgList,
+    pub timeout: u64, // in seconds
 }
 
 impl Default for Config {
@@ -15,6 +16,7 @@ impl Default for Config {
             algs: algorithm::AlgList::client_default(),
             auth: auth::AuthInfo::default(),
             ver: version::SshVersion::default(),
+            timeout: 30,
         }
     }
 }
@@ -26,6 +28,7 @@ impl Config {
             algs: algorithm::AlgList::default(),
             auth: auth::AuthInfo::default(),
             ver: version::SshVersion::default(),
+            timeout: 30,
         }
     }
 }
